refactor(attention): move SFA CP context types to common_cp
Relocate DCPQueryGatherContext, DCPContext, and DSACPContext from
sfa_v1 into context_parallel/common_cp so shared CP metadata lives
together and sfa_cp no longer imports these types from the mainline
SFA module.
